Move subtract fee calculation into transaction mod
